Excellent trick to solve Jupiter Attacks with a modified Fenwick Tree.

+++ ICPC Live Archive/5798 - Jupiter Attacks/check.rb
@@ -1,13 +1,18 @@
-10.times do |i|
+20.times do |i|
   puts "Test #{i}..."
   
   system "ruby gen.rb > rand.in"
   system "./bruteforce < rand.in > right.out"
-  system "./jupiter < rand.in > wrong.out"
+  exit -1 unless $?.success?
+  system "./jupiter.4 < rand.in > wrong.out"
   system "diff right.out wrong.out"
   unless $?.success?
+    puts
+    puts "*" * 10
+    puts 
+    system "cat rand.in"
+    system "rm right.out wrong.out"    
     exit -1
-    system "rm right.out wrong.out"
   end
 end
 system "rm right.out wrong.out"
